+++ hpOneView/resources/networking/sas_interconnect_types.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -30,10 +30,10 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import Resource, unavailable_method
 
 
-class SasInterconnectTypes(object):
+class SasInterconnectTypes(Resource):
     """
     SAS Interconnect Types API client.
 
@@ -43,58 +43,17 @@ class SasInterconnectTypes(object):
     """
     URI = '/rest/sas-interconnect-types'
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(SasInterconnectTypes, self).__init__(connection, data)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets a paginated collection of SAS interconnect types. The collection is based on optional
-        sorting and filtering and is constrained by start and count parameters.
+    def create(self):
+        """Create method is not available"""
+        unavailable_method()
 
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
+    def delete(self):
+        """Delete method is not avaialble"""
+        unavailable_method()
 
-        Returns:
-            list: A list of SAS interconnect types.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get(self, id_or_uri):
-        """
-        Gets the SAS interconnect type with the specified ID or URI.
-
-        Args:
-            id_or_uri: ID or URI of the SAS Interconnect Type.
-
-        Returns:
-            dict: The SAS Interconnect Type
-        """
-        return self._client.get(id_or_uri)
-
-    def get_by(self, field, value):
-        """
-        Gets all SAS interconnect types that match the filter.
-
-        The search is case-insensitive.
-
-        Args:
-            field: Field name to filter.
-            value: Value to filter.
-
-        Returns:
-            list: A list of SAS interconnect types.
-        """
-        return self._client.get_by(field, value)
+    def update(self):
+        """Update method is not avaialble"""
+        unavailable_method()